Implement TttInferenceClient and TttInferenceServer
